+++ test/smoke/pagerduty/pagerduty_test.rb
@@ -15,11 +15,13 @@ if os.redhat?
   pagerduty_cgi_dir = '/usr/lib64/nagios/cgi-bin'
   path_config_dir   = '/etc/nagios/conf.d'
   perl_cgi_package  = 'perl-CGI'
+  plugin_dir        = '/usr/lib64/nagios/plugins'
 else
   command_file      = "/var/lib/#{vname}/rw/nagios.cmd"
   pagerduty_cgi_dir = "/usr/lib/cgi-bin/#{vname}"
   path_config_dir   = "/etc/#{vname}/conf.d"
   perl_cgi_package  = 'libcgi-pm-perl'
+  plugin_dir        = '/usr/lib/nagios/plugins'
 end
 
 # PagerDuty Configuration
@@ -37,6 +39,11 @@ describe package(perl_cgi_package) do
 end
 
 # Test Pagerduty Integration Script
+
+describe command "perl #{plugin_dir}/notify_pagerduty.pl" do
+  its('stderr') { should match /pagerduty_nagios enqueue/ }
+  its('exit_status') { should eq 2 }
+end
 
 describe file("#{pagerduty_cgi_dir}/pagerduty.cgi") do
   its(:content) { should match "'command_file' => '#{command_file}'" }